requirements.txt
asgiref==3.7.2
Django==4.2.4
django-coverage-plugin==3.1.0
django-mysql==4.11.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
iniconfig==2.0.0
mixer==7.2.2
mysql-connector-python==8.1.0
packaging==23.1
pluggy==1.2.0
protobuf==4.21.12
pyreadline3==3.4.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
